--- src/components/NarrowView.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function NarrowView() {
-    return <div />
-}
