+++ src/components/NarrowView/SettingsView/ScaleControl.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Control, { ControlProps } from '../Control'
-import ScaleSelect from '../ScaleSelect'
+import Control, { ControlProps } from '../../Control'
+import ScaleSelect from '../../ScaleSelect'
 
 export default function ScaleControl({ label = 'Scale', ...props }: ControlProps) {
     return (
