+++ src/features/view/types.ts
@@ -1,6 +1,6 @@
-import SettingsView from '../../components/SettingsView'
-import MetricsView from '../../components/MetricsView'
-import ProfilesView from '../../components/ProfilesView'
+import SettingsView from '../../components/NarrowView/SettingsView'
+import MetricsView from '../../components/NarrowView/MetricView'
+import ProfilesView from '../../components/NarrowView/ProfilesView'
 import SettingsIcon from '../../icons/SettingsIcon'
 import MetricsIcon from '../../icons/MetricsIcon'
 import ProfilesIcon from '../../icons/ProfilesIcon'
