+++ src/utils/mlToL.ts
@@ -1,0 +1,3 @@
+export default function mlToL(value: number) {
+    return (value / 1000).toFixed(1)
+}
